CSN: CSN-m5-48
Switch to m5-stable from 2010/10/29 with local patches.

Bug fixes:
  main.py checked the wrong verbose flag.
  system.hh should use 64 bits for the page_ptr (fails with large memory).

Local changes for exporting exit() value to HAsim and for memory read/write
callback hooks.

+++ m5-releases/20101029-f299139501f7-stable/src/sim/system.hh
@@ -136,7 +136,7 @@ class System : public SimObject
 
 #else
 
-    int page_ptr;
+    uint64_t page_ptr;
 
   protected:
     uint64_t next_PID;
